pkg/process: Fix more tests that should not run on windows
