+++ pkg/process/checks/allprocesses_windows_test.go
@@ -54,7 +54,7 @@ func TestCommandLineSplitting(t *testing.T) {
 			},
 		},
 	} {
-		assert.Equal(t, tc.expected, tc.input)
+		assert.Equal(t, tc.expected, parseCmdLineArgs(tc.input))
 	}
 }
