+++ pkg/process/net/uds_test.go
@@ -1,3 +1,5 @@
+// +build linux
+
 package net
 
 import (
